stream-service: Restart accepting without blocking
Calling on_accept() sometimes lead to deadlocks when service->destroy()
was called concurrently.  That is, two threads waiting in on_accept() but
the last worker would only wake one due to the call to signal().  Calling
broadcast() wouldn't help either as that could lead to crashes if the thread
that called destroy() is waked first.

This also makes this more efficient as a constant pool of concurrent workers
can be maintained, otherwise peaks at the limit were followed by only a single
worker being active.